backend/requirements.txt
flask==3.1.1
gevent==25.5.1
psycopg2-binary==2.9.10
redis==6.2.0
SQLAlchemy==2.0.41
pydantic==2.11.5
structlog==25.4.0
rich==14.0.0
textual==3.3.0
click==8.2.1
dependency-injector==4.47.1
pytest==8.3.5
pytest-asyncio==1.0.0